fix build.sh

+++ build.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
 
-rm output.lua
+if [ -f output.lua ]
+then
+  rm output.lua
+fi
 
 process_file() {
   cat ../ARC1-Extensions/src/ARC1-$1.lua >> output.lua
 }
 
 process_file "Core"
-#process_file "Burnable"
-#process_file "Mintable"
+# "Burnable" and "Mintable" SHOULD NOT be included!
 process_file "Pausable"
 process_file "Blacklist"
 process_file "allApproval"
